Create SpreadCardsContainer
