+++ src/containers/index.js
@@ -2,3 +2,4 @@ export { default as AppContainer } from './AppContainer'
 export { default as CardContainer } from './CardContainer'
 export { default as CardDetailsContainer } from './CardDetailsContainer'
 export { default as HeaderContainer } from './HeaderContainer'
+export { default as SpreadCardsContainer } from './SpreadCardsContainer'
